ProjectUploadTask modified to handle JSONObjects, answers from server now displaying in GUI

+++ catroid/src/at/tugraz/ist/catroid/constructionSite/tasks/ProjectUploadTask.java
@@ -5,6 +5,9 @@ import java.io.FilenameFilter;
 import java.io.IOException;
 import java.util.HashMap;
 
+import org.json.JSONException;
+import org.json.JSONObject;
+
 import android.app.ProgressDialog;
 import android.app.AlertDialog.Builder;
 import android.content.Context;
@@ -18,7 +21,7 @@ import at.tugraz.ist.catroid.web.WebconnectionException;
 public class ProjectUploadTask extends AsyncTask<Void, Void, Boolean> {
 	private static final String FILE_UPLOAD_TAG = "upload";
 	private static final String PROJECT_NAME_TAG = "projectTitle";
-	private static final String FILE_UPLOAD_URL = "http://www.catroid.org/catroid/upload/upload.http";
+	private static final String FILE_UPLOAD_URL = "http://www.catroid.org/catroid/upload/upload.json";
 	private static final String TEST_FILE_UPLOAD_URL = "http://catroidwebtest.ist.tugraz.at/catroid/upload/upload.http";
 	
 	public static boolean mUseTestUrl = false;
@@ -29,6 +32,7 @@ public class ProjectUploadTask extends AsyncTask<Void, Void, Boolean> {
 	private ProgressDialog mProgressdialog;
 	private String mProjectName;
 	private String resultString;
+	private String mServerAnswer;
 	
 	// mock object testing
 	protected ConnectionWrapper createConnection() {
@@ -40,6 +44,7 @@ public class ProjectUploadTask extends AsyncTask<Void, Void, Boolean> {
 		mProjectPath = projectPath;
 		mZipFile = zipFile;
 		mProjectName = projectName;
+		mServerAnswer = "An error occurred while uploading the project.";
 	}
 	
 	@Override
@@ -90,8 +95,29 @@ public class ProjectUploadTask extends AsyncTask<Void, Void, Boolean> {
 			else
 				serverUrl = FILE_UPLOAD_URL;
 			resultString = createConnection().doHttpPostFileUpload(serverUrl, hm, FILE_UPLOAD_TAG, mZipFile);
-			file.delete();
-			return true;
+			
+			
+			JSONObject jsonObject = null;
+			int statusCode = 0;
+			
+			try {
+				jsonObject = new JSONObject(resultString);
+				statusCode = jsonObject.getInt("statusCode");
+				mServerAnswer = jsonObject.getString("answer");
+			} catch (JSONException e) {
+				e.printStackTrace();
+			}
+			if (statusCode == 200) {
+				file.delete();
+				return true;
+			}
+			else if (statusCode >= 500) {
+				return false;
+			}
+			
+			return false;
+			
+			
 		} catch (IOException e) {
 			e.printStackTrace();
 		} catch (WebconnectionException e) {
@@ -108,19 +134,19 @@ public class ProjectUploadTask extends AsyncTask<Void, Void, Boolean> {
 			mProgressdialog.dismiss();
 		
 		if(!result) {
-			showDialog(R.string.error_project_upload); 
+			showDialog(mServerAnswer); 
 			return;
 		}
 		
-		showDialog(R.string.success_project_upload);
+		showDialog(mServerAnswer);
 		
 	}
 	
-	private void showDialog(int messageId) {
+	private void showDialog(String message) {
 		if(mContext == null)
 			return;
 		new Builder(mContext)
-			.setMessage(messageId)
+			.setMessage(message)
 			.setPositiveButton("OK", null)
 			.show();
 	}
